requirements.txt
flask
pyrebase4
pycryptodome
gunicorn
